langchain/openai-translator/requirements.txt
pdfplumber
simplejson
requests
PyYAML
pillow
reportlab
pandas
loguru
openai
langchain
gradio
flask
langchain_openai
zhipuai
langchain_community